Updated duplicate detection and solving
Implemented some functionality to detect duplicates! Will now find if there is matching masking data and, if so, uses the PSF data from that night. If there is no masking data, it will use the entry with the highest magnitude contrast since that means we detected a fainter companion.

The plot_data function is currently broken. This will be fixed in the next commit.