Loads of new toys!
Edited data_parser.py so it loops through the plots. I have a feeling I will be exporting the final plot to a PDF soon enough (perhaps with the next version!).

Added praesepe_cmd_mass.py. Currently, this is simply a copy of a file I used to practice plotting a couple of weeks back. Eventually, I will be updating it to get the masses of the targeted objects.

Added bhac_gaia_2mass_reader.py. Currently, this is a script to read in the GAIA and 2MASS BHAC data for a given age. With the next version, I might change this to a function so it can be easily changed.