Intermediate updates
Added loads of CSV files of varying usefulness. Most will be deleted with the next commit.

Used data_parser.py and bhac_gaia_2mass_reader.py to create them (hence most of their changes). The former was also changed to output the different designations of our targets to a single CSV.

Now have the masses of the stars using color and absolute magnitude and plotted them. Fixing up this plot is the next step! :)