Important module_runner revamp
targets_abr.csv was incorrectly created. The data in each row did not match the appropriate target star. module_runner.py is in the process of being changed to correct this. Checks to ensure that all of the data is correctly paired to the appropriate star have been implemented, as well.

targets_abr2.csv is temporarily serving as the correct table. Once all of the necessary information has been transferred to this table, it will overwrite targets_abr.csv (the original, currently inaccurate, table).