Miscellaneous updates
Added RA and DEC columns to our targets table. The functionality to do this was implemented in module_runner.py.

HSHJ510 has been removed from our targets list since it is not a member of Praesepe.

I found HIP14807, which we have masking and PSF data for. It turns out to be left over from a former collaborator!